requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
python-docx>=0.8.11 